ci: Publish artifacts to build.l5d.io
In order to setup continuous integration, proxy artifacts need to be
published somewhere predictable and discoverable. This change configures
Travis CI to publish proxy artifacts built from master to:

    build.l5d.io/linkerd2-proxy/linkerd2-proxy-${ref}.tar.gz
    build.l5d.io/linkerd2-proxy/linkerd2-proxy-${ref}.txt

The tarball includes an optimized proxy binary and metadata (right now, just
the LICENSE file, but later this should include additional version/build
metadata that can be used for runtime diagnostics).

The text file includes the sha256 sum of the tarball.

A `Makefile` is introduced to encapsulate build logic so that it can both
drive CI and be used manually.

Travis CI is configured to run debug-mode tests against PRs and to run a full
release package-test-publish for commits to
master.