refactor: Make `Background::query_destination_service_if_relevant` into a method (#35)
This is strictly a refactor which should make no functional changes.

Currently, the function used to construct new Destination service
queries (`Background::query_destination_service_if_relevant`) is a
function rather than a method on `Background`, although it takes as an
argument a field from `Background`. This is because in some cases, it is
called where `self.destinations` is borrowed mutably, preventing `self`
from being borrowed immutably.

Right now, this means that one additional field has to be passed
explicitly. However, in order to add the limit on active Destination
service queries, it was necessary to add two additional fields to
`Background` that have to be passed to this function. Since these
arguments should always come from fields on `Background`, it would be
preferable for this to be a method.

This branch breaks out some of the fields on
`control::destination::Background` into their own structs:
`DestinationCache`, which holds the map of DNS names to
`DestinationSet`s and the queue of DNS names that need reconnects; and
`Config`, which holds the configuration necessary to create a new
Destination service query (currently just the `Namespaces` config). This
allows us to have separate borrows on the `DestinationCache` and
`Config`. 

When I make the additional changes necessary to add the limit on active 
destination queries, the two additional fields necessary can be added to 
`Confiq`, rather than having to explicitly pass them into 
`query_destination_service_if_relevant` every time it's called.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>