Split tls_config_reload into read and write halves
Rather than exposing a single, shared, `Arc<TlsConfigReload>` type, this
change splits the concerns between a `Sensor` type and a `Fmt` type.

The `Sensor` type holds a strong reference to the inner metrics
structure and acquires a write lock on updates.

The `Fmt` type holds a weak reference to the inner metrics structure so
that the metrics server can print the metrics state, as long as it's
actually held in memory.  If the `Sensor` is dropped (for instance,
because TLS has been administratively disabled), then no metrics will be
formatted by `Fmt`.