Remove `record` benches
As we begin to remove event-based metrics telemetry, the `record` bench
test becomes an obstacle that needs to be updated at every step.

Given that we have already removed the event channel and plan to remove
the `Record` type entirely, it's suitable to remove the bench test
entirely.

While this is done, the `test-benches` Make target and rust-nightly
travis stage are obsoleted.