Hide HTTP telemetry implementation details
Previously, much of `telemetry::http`'s types and internal
implementation details are exposed to the rest of the telemetry system.
In preparation for further changes to support more granular locking,
this change makes metric storage and recording implementation details.

Following this, `telemetry::http` exposes a `Report` type for printing
metrics to the server and a `Sensors` type used to instrument stacks
with HTTP telemetry. These types share an internally-mutable metrics
registry that is private to the http module.

The `event` types continue to be exposed to support Tap, but the
convenience exports have been removed.

The `metrics::Root` type no longer needs to be shareable. This type will
be replaced in a followup change.