Introduce svc::Layer for building stacks
`proxy::h2_router` is currently responsible for initilaizing the
`TimestampRequestOpen` middleware, but this has nothing to do with
routing.

In order to separate these concerns but maintain the `svc::MakeClient`
API expected by `proxy::Server`, this change introduces `svc::Layer` so
that the `TimestampRequestOpen` middleware can easily be made to wrap
`h2_router::Make` to produce a nwe `MakeClient`.