Split proxy infrastructure from `bind`
The `bind` module includes core types that are used throughout the proxy
module, like `bind::Protocol`. Furthermore, this protocol is not really
related to the protocol detection that occurs in `proxy`.

In order to simplify `bind` and set up for further refactoring (adopting
`MakeClient`), this change reorganizes the module structure as follows:

- `bind::Protocol` is now `proxy::http::Dialect`. This change is made to
   distinguish this concept from protocol detection and, furthermore, to
   separate this logic from endpoint-construction implemented in `bind`.

- HTTP proxying utilities have been moved into the `proxy::http` module.

- `bind::NormalizeUri` is now `proxy::http::NormalizeUri`

- `proxy::h2_router` is now `proxy::http::router`