Remove settings checks from orig-proto module
When constructing the `outbound::orig_proto_upgrade` module, we check
the endpoint's `settings` to determine the `Upgrade` middleware should
be applied. The `Upgrade` middleware performs a similar, and partially
redundant set of checks on each request.

linkerd/linkerd2#1798 is simplif if we can separate destination-routing
from client settings. Specifically, if the `settings` field will be
removed from `Destination`s and `Endpoint`s, we won't a discovery &
routing stack for each combination of Settings. Instead, a routing layer
will be installed around the client, beneath the profile route stack,
that is responsible for dispatching requests to the appropriate client.

This change does NOT change user-facing behavior. An additional
`h1::wants_upgrade` check is now perormed within the `Upgrade` service
so that HTTP-Upgrade requests can pass through this stack.