Implement Error for svc::Either
The `linkerd2_stack::Either` type is used to implement Layer, Stack, and
Service for alternate underlying implementations. However, the Service
implementation requires that both inner service emit the same type of
Error.

In order to allow the underlying type to emit different errors, this
chnage uses `Either` to wrap the underlying errors, and implements
`Error` for `Either`.