profiles: Drive profile discovery on a daemon task
The profile router currently is responsible for driving the state of
profile discovery; but this means that, if a service is not polled for
traffic, the proxy may not drive discovery (so that requests may
timeout, etc).

This change moves this discovery onto a daemon task that sends profile
updates to the service over an mpsc with capacity of 1.