Actually fix the master CI build (#164)
It turns out that increasing the recursion limit for the `tap` test
crate _actually_ fixes the compiler error that's broken the last several
builds on master. 

Since I'm now able to locally reproduce the error (which occurs only
when running the tests in release mode), I've verified that this
actually does fix the issue. Thus, I've also reverted the previous
commit (7c35f27ad3e86614c05a2a409c3a8b459cf023ac) which claimed to fix
this issue, as it turns out that was not actually necessary.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>