Make timeouts for canonicalization DNS queries tuneable (#175)
Currently, the proxy hard-codes a 100ms timeout for DNS queries to
canonicalize a name. In some cases (see linkerd/linkerd2#2069), a DNS
server might be slow enough to respond that a majority of these queries
exceed the timeout. This results in a lot of error messages being
logged, and potentially, a delay or complete failure to update to
changes in DNS, if the DNS server is consistently very slow.

This branch adds a `LINKERD2_PROXY_DNS_CANONICALIZE_TIMEOUT` environment
variable that allows the default 100ms timeout to be overridden. 

See also linkerd/linkerd2#2093

Signed-off-by: Eliza Weisman <eliza@buoyant.io>