canonicalize: Only log errors at the WARN level when falling back (#174)
Previously, whenever the `canonicalize::Task` future encounters an
error, it logs that error at the error level. However, in many cases,
these errors are transient, and we are able to successfully fall back to
a previously canonicalized name.

This branch changes `canonicalize::Task` to only log at the error level
when there's no previously successful result to fall back to, and log at
the warning level otherwise. In addition, the log message on fallbacks
now indicates that we fell back to a previous canonicalization.

Hopefully, this should make transient errors, such as a slow DNS server,
a little less scary.

Fixes linkerd/linkerd2#2094

Signed-off-by: Eliza Weisman <eliza@buoyant.io>