Remove inotify dependency from the main proxy crate
AFAICT the main proxy crate's inotify dependency is not actually used,
as the `fs-watch` lib crate fully encapsulates the dependency. I think
that this was overlooked when factoring out `fs-watch` into its own
crate.

It seems better to not specify the same dependency in multiple places,
especially since we're not using one of them...

Signed-off-by: Eliza Weisman <eliza@buoyant.io>