Fix inotify dependency (#185)
Since my PR against `inotify` (inotify-rs/inotify#124) has merged, we can
go back to depending on the upstream version. Also, this updates the SHA
we depend on to one that actually exists, which should fix the build.

I've also removed the `inotify` dependency from the main proxy crate's
`Cargo.toml.` This dependency is not actually used, as the `fs-watch` lib
crate fully encapsulates the dependency. I think that this was overlooked
when factoring out `fs-watch` into its own crate.

It seems better to not specify the same dependency in multiple places,
especially since we're not using one of them.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>