Set proxy_id in calls to Get and GetProfile (#183)
Pick up API changes to the proxy-api where Get and GetProfiles now accept a proxy_id parameter.  We set this parameter to be equal to the proxy's TLS pod identity, which is configured through an environment variable.  Sending this value to the proxy-api allows it to return more tailored results, for example, by looking for service profiles in the proxy's namespace.

Signed-off-by: Alex Leong <alex@buoyant.io>