Remove the CONTROLLER_NAMESPACE configuration
linkerd/linkerd#2371 removes the need for the proxy to be aware of the
controller namespace by moving the TLS-identity checking logic into the
destination service.

With this change, when the destination service returns an identity, the
proxy will try to establish TLS'd communication to the endpoint if the
proxy has trust anchors.