perf(router): Increase inbound/outbound router cap.
In both server and client case, the default value of
MAX_CONCURRENT_STREAMS is `u32::MAX`. However, we currently cap the
inbound and outbound router capacity limits to 100.

This change increases the capacity of both to 10_000.

This will Close linkerd/linkerd2#2395 and could possibly address the
issue in linkerd/linkerd2#2363.

Signed-off-by: Kevin Leimkuhler <kevinl@buoyant.io>