Instrument identity with logging (#217)
This changes initialization logic so that a task is spawn that logs when
identity is provisioned (as observed by stacks). Identity's daemon task is
now spawned on the admin thread with context logging.

Furthermore, the processes's uptime is now logged in log messages so that the
relative time between events is easily observeable. This is borrowed from the
Linux kernel's behavior (i.e. in /var/log/kern.log).