Add test support infrastructure for identity (#227)
This branch adds a mock Identity service implementation to the proxy's
test-support module, and wires up the rest of the test support code to
allow mocking the Identity service in tests. The existing tests for the
`l5d-server-id` header have been updated to work with the identity
service.

Note that this branch does *not* add any new tests for the proxy's
behaviour with the Identity service. I was planning to add these in a
follow-up PR.

See also linkerd/linkerd2#2505

Signed-off-by: Eliza Weisma <eliza@buoyant.io>