Rewrite the destination client and remove DNS fallback (#259)
This branch removes the fall-back-to-DNS behaviour from the proxy's
destination module. Instead, when the Destination service returns "no
endpoints", the proxy will route the request to its original
destination, using the fallback behaviour introduced in #248.

In addition, the proxy now no longer reconnects to the destination
service for queries which receive a response with the Invalid Argument
status code. After linkerd/linkerd2#2664, this status is used by the
control plane to indicate names which should not query the destination
service. I've added a new test for this.

Furthermore (and unlike #245), this branch also removes the caching of
service discovery lookups so that they can be shared between resolvers
for authorities with the same canonical form. When we limited the number
of concurrently in flight destination service queries to 100, reusing
them was much more important; now it adds a lot of unnecessary
complexity.

The rewritten `control::destination` module is now much simpler and
(hopefully) easier to understand. The `Resolver` holds a destination
service `Client`, which is cloned every time a new `Resolution` is
requested. Each active `Resolution` has a corresponding background task
that drives the destination service query and sends updates through a
channel to the `Resolver`. 

The background task is necessary so that updates are processed as they
are recieved, rather than only when the `Resolution` is polled (i.e.
when the load balancer is routing a request). In addition, the channel
provides the queueing necessary to transform the control plane's update
types (which contain multiple endpoint addresses) into multiple
`tower-balance` updates for single endpoints, and to invalidate stale
endpoints when the control plane connection is reestablished. When a
control plane query is not necessary because the requested authority
does not match the proxy's destination suffixes, the background task is
not spawned and the `Resolution` remains permanently in the "no
endpoints" state; and when an `InvalidArgument` error code is recieved,
the background task terminates.

A majority of the service discovery tests pass with no changes after
this change. The following changes were necessary: 
* The test `outbound_updates_newer_services` was removed, as it tests 
  that the same destination service query is shared between multiple
  services with the same authority (which is no longer the case).
* The mock destination service was changed to send the `Unavailable`
  status code when it has run out of expected destinations, rather than
  `InvalidArgument`, as that code is used to indicate that a destination
  should never be reconnected. 
* New tests were added for falling back to ORIG_DST when no endpoints 
  exist, and for not reconnecting after an `InvalidArgument` error.

Closes #245.
Closes linkerd/linkerd2#2661
Closes linkerd/linkerd2#2728

Signed-off-by: Eliza Weisman eliza@buoyant.io