Use a constant-time load balancer
tower-rs/tower#288 changed the Balance implementation substantially so
that it is no longer responsible for driving services to readiness.
Instead, a new layer, `spawn_ready` is inserted around endpoint stack to
ensure that readiness is driven on a background task.

In support of this change, the `pending` layer was removed from the
enpdoint stack and, instead, the discovery system is now responsible for
driving pending services to be materialized.