metrics: Include the prefix of a Report in log lines
This branch changes the `proxy::http_metrics::Report` type so that when
the metrics report has a prefix (such as "route" or "control", etc), the
prefix is included in log lines emitted by the report.

Since the log line produced by `Report` includes the number of entries
in the metrics registry for that `Report`, this should make it much
easier to see which metrics scope that count corresponds to.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>