Honor `l5d-override-dst` for inbound service profiles (#267)
When the proxy receives an inbound request, it strips the
`l5d-dst-override` header and ignores its value. Instead, this header
value should be honored in the inbound router as it is in the outbound
router. This will enable service profile discovery from ingresses, for
instance.

This branch changes the inbound router to use the `l5d-override-dst`
header preferentially, with only the `l5d-canonical-dst` header taking
priority over it if it is present. The layer that strips the override
header is moved to under the inbound dst router.

In addition, I've added new tests in the discovery module that assert
that the value of the `l5d-override-dst` header is used for profile
discovery when present, on both the inbound and outbound sides of the
proxy. The inbound tests fail prior to the other changes on this branch.

Closes linkerd/linkerd2#2910.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>