Introduce `l5d-require-id` request header (#289)
This was originally implemented in #280, but has since changed enough that a
new PR is being opened

### Motivation

As part of the Tap hardening project, tap should only be served on TLS
connections. Currently, when the Tap service makes an outbound Tap request, the
`Host` header is the Pod IP of the resource. Service discovery is not attempted
and therefore the Tap service initiates a plaintext connection. The introduction
of the `l5d-require-id` header will allow an outbound request to require the
specified identity on the endpoint it connects to.

### Details

The `orig_dst_router` router on outbound requests will set the identity of the
`outbound::Endpoint` to the value of the `l5d-require-id` header if it is
present. This is useful specifically for the case explained above when request
are made with an IP as their `Host` header and need to be TLS connections.

In order for the header to be used in other scenarios, a new layer has been
added to the `endpoint_stack` stack on outbound requests. This layer is
implemented in `outbound::require_identity_on_endpoint`. The purpose of this
layer is to wrap the `MakeService` of `Endpoint`s with a new service that checks
for the presence of required identities, and if present, asserts the identity of
the endpoint is expected.

When constructing an `Endpoint`, if a request has the `l5d-require-id` header
then the endpoint will be constructed if it's identity is the expected identity
of the header value; if not, it will error and return a status code `FORBIDDEN`.

### Tests

Tests have been added for both http1 and http2. The tests cover the
`orig_dst_router` router as well as discovery. In both of those cases, clients
make requests with and without the `l5d-require-id` header, as well as correct
and inccorect `l5d-require-id` header values.

Closes linkerd/linkerd2#2677

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>