Tap server authorizes clients when identity is expected (#290)
### Motivation

As part of the tap hardening project the tap server must serve tap requests
exclusively to the tap service. Clients must be both authenticated and
authorized. Currently, the tap server still can accept clients over plaintext,
and does not authorize clients over private connections.


### Details

In order for authenticated clients to be authorized, the `tap_identity` config
variable has been introduced. This value is used as the expected client identity
of incoming tap requests. When `tap_identity` is set, the tap server now checks
the peer identity of the client connection against this value. If they are equal
the connection is accept; otherwise is is rejected.

The tap server still accepts plaintext connections. This will be removed in a
future PR once the tap controller is sending tap requests with the required
`l5d-require-id` that is needed to set server identity on outbound requests to
IP addresses.

### Tests

Tests have been added to `tap::expected_identity` that assert:

* The tap server accepts plaintext connections when `tap_identity` is not set in
  the config
* The tap server rejects plaintext connections when `tap_identity` is set in the
  config
* The tap server rejects private connections when the client identity does not
  match the expected `tap_identity`
* The tap server accepts private connections when the client identity does match
  the expected `tap_identity`

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>