tap: fix tap response streams never ending (#299)
A bug currently exists in the tap code where response streams never end
when reaching the limit on the number of events. Since these streams
never end, the tap controller never restarts them, causing the
user-facing tap output to hang.

This is due to a bug introduced in
073a1beb4a7cd709c6b1eaa56a319c1829a94d11. The `Dispatch` future
previously was responsible for holding a sender to the events channel,
and cloning it into each tapped service. In that change, the `Dispatch`
type was removed, and the cloned sender was moved to the `Tap` struct.
The `Dispatch` future was dropped when the limit is reached, but the
`Tap` struct is not. This means that the final sender is never dropped,
and the events stream never ends.

This commit fixes the bug by moving the sender into `Shared`, which
`Tap` holds a `Weak` reference to. This ensures that the sender that's
cloned into each tapped service is dropped when the limit is reached,
allowing the channel to close when all other senders are dropped.

I've manually verified that the hang as described in
linkerd/linkerd2#2891 no longer occurs after making this change.

Fixes linkerd/linkerd2#2891

Signed-off-by: Eliza Weisman <eliza@buoyant.io>