Authority label should reflect logical dst (#300)
Supersedes https://github.com/linkerd/linkerd2-proxy/pull/296

The authority label reflects the concrete dst of the request in the endpoint stack.  In the case of a traffic split, this may be different from the actual authority of the request (i.e. the logical dst).

We add the logical dst as a field on the endpoint target and use this value for the authority label instead of the concrete dst.

The net result is that the authority field reflects the apex service of a traffic split and the `dst_service` label reflects the leaf service.  

Signed-off-by: Alex Leong <alex@buoyant.io>