core: Formalize the listen/serve API
The `transport::tls::Listen` and `proxy::Server` types together formed a
sort of interface that facilitiates serving connections from tasks.

This change formalizes these interfaces so that the concrete
implementations can move/change more easily.

In pursuing this change, the remote `SocketAddr` is no longer passed
around in a tuple and is, instead, wired into the `Connection` type. The
`AddrInfo` trait has been updated to account for existince of the remote
socket addr on a socket.