Introduce linkerd2-proxy-core
The linkerd2-proxy-core crate is intended to expose the core types and
interfaces needed to implement linkerd2-proxy so that implementation
pieces may eventually be split into separate sub-crates.

Initially, the crate exposes only `Error`, which is redefined
gratuitously throughout the project; and it re-exports the commonly used
`Never` type.

Additional traits and core types may be added in follow-up changes.

The proxy has been updated to re-export many of the core types and
modules from the root of the `linkerd2-proxy` crate and imports have
been updated accordingly.

No functional changes.