fallback: Split fallback into dedicated crate
The `proxy::http::fallback` module need not be in the main proxy crate.
Furthermore, it shouldn't depend on the `svc::Builder` type, which
really just holds a `Layer`. By upgrading the tower dependency, we can
extract the inner layer from the builder and make the `fallback::Layer`
type implemented in terms of these layers.